test: migrate app-action-call tests
